requirements.txt
cycler==0.10.0
dill==0.2.9
kiwisolver==1.1.0
matplotlib==3.0.3
numpy==1.16.3
pkg-resources==0.0.0
progressbar2==3.39.3
pyparsing==2.4.0
python-dateutil==2.8.0
python-utils==2.3.0
scipy==1.2.1
six==1.12.0
